[REFACTOR] 변수 수정

+++ DDIS_HR/src/main/java/com/ddis/ddis_hr/member/command/application/dto/EmployeeDTO.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @ToString
 public class EmployeeDTO {
 
-    private Integer employeeId;
+    private Long employeeId;
     private String employeeName;
     private String employeePwd;
     private String employeePhotoName;
